// src/app/config/shedule/interfaces/days.interface.ts
type ITeacher =
  | 'Богдашич Елена Ефимовна'
  | 'Антонова Анна Владимировна'
  | 'Григорьева Анна Олеговна'
  | 'Бондаренко Павел Игоревич'
  | 'Марков Владислав Игоревич'
  | 'Богдан Светлана Ивановна'
  | 'Прохорчук Анна Владимировна'
  | 'Кривицкий Сергей Валентинович'
  | 'Праволоцкий Алексей Николаевич'
  | 'Антипенко Татьяна Игоревна';

export interface ILecture {
  name:
    | 'Физра'
    | 'Охрана среды'
    | 'Англ яз'
    | 'Пд и УП'
    | 'Коммуникации'
    | 'ИР ВР'
    | 'Сети'
    | 'Безопасность'
    | 'ООС и ЭС'
    | 'Права';
  timeDuration: '8:00-9:40' | '9:55-11:35' | '12:15-13:55' | '14:10-15:50' | '16:20-18:00' | '18:15-19:55';
  teachers: ITeacher[];
  audition: '319' | '302' | '215' | '302' | '307' | '323' | 'Спортзал' | '118' | '205' | '318' | '309';
  type: 'lecture';
}

interface ILunch {
  timeDuration: '11:35-12:15' | '15:50-16:20';
  type: 'lunch';
  name: 'Обед';
}

export type IDayEvent = ILecture | ILunch;

export interface IDay {
  name: 'Понедельник' | 'Вторник' | 'Среда' | 'Четверг' | 'Пятница' | 'Суббота';
  duration?: string;
  events: IDayEvent[];
}